+++ example/lib/fixed_example.dart
@@ -1,0 +1,86 @@
+import 'package:drag_and_drop_lists/drag_and_drop_item.dart';
+import 'package:drag_and_drop_lists/drag_and_drop_lists.dart';
+import 'package:example/navigation_drawer.dart';
+import 'package:flutter/material.dart';
+import 'package:flutter/widgets.dart';
+
+class FixedExample extends StatefulWidget {
+  FixedExample({Key key, this.title}) : super(key: key);
+  final String title;
+
+  @override
+  _FixedExample createState() => _FixedExample();
+}
+
+class _FixedExample extends State<FixedExample> {
+  List<DragAndDropList> _contents;
+
+  @override
+  void initState() {
+    super.initState();
+
+    _contents = List.generate(10, (index) {
+      return DragAndDropList(
+        header: Row(
+          children: <Widget>[
+            Expanded(
+              flex: 1,
+              child: Divider(),
+            ),
+            Padding(
+              padding: const EdgeInsets.symmetric(horizontal: 10, vertical: 5),
+              child: Text(index % 4 == 0 ? 'Header $index : Non-Draggable' : 'Header $index'),
+            ),
+            Expanded(
+              flex: 1,
+              child: Divider(),
+            ),
+          ],
+        ),
+        children: <DragAndDropItem>[
+          DragAndDropItem(
+            child: Text('$index.1'),
+          ),
+          DragAndDropItem(
+            child: Text('$index.2 : Non-Draggable'),
+            canDrag: false,
+          ),
+          DragAndDropItem(
+            child: Text('$index.3'),
+          ),
+        ],
+        canDrag: index % 4 != 0,
+      );
+    });
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        title: Text('Basic'),
+      ),
+      drawer: NavigationDrawer(),
+      body: DragAndDropLists(
+        children: _contents,
+        onItemReorder: _onItemReorder,
+        onListReorder: _onListReorder,
+      ),
+    );
+  }
+
+  _onItemReorder(
+      int oldItemIndex, int oldListIndex, int newItemIndex, int newListIndex) {
+    setState(() {
+      var movedItem = _contents[oldListIndex].children.removeAt(oldItemIndex);
+      _contents[newListIndex].children.insert(newItemIndex, movedItem);
+    });
+  }
+
+  _onListReorder(int oldListIndex, int newListIndex) {
+    setState(() {
+      var movedList = _contents.removeAt(oldListIndex);
+      _contents.insert(newListIndex, movedList);
+    });
+  }
+}