+++ example/lib/main.dart
@@ -1,6 +1,7 @@
 import 'package:example/basic_example.dart';
 import 'package:example/drag_into_list_example.dart';
 import 'package:example/expansion_tile_example.dart';
+import 'package:example/fixed_example.dart';
 import 'package:example/horizontal_example.dart';
 import 'package:example/list_tile_example.dart';
 import 'package:example/sliver_example.dart';
@@ -28,6 +29,7 @@ class MyApp extends StatelessWidget {
         '/sliver_example': (context) => SliverExample(),
         '/horizontal_example': (context) => HorizontalExample(),
         '/drag_into_list_example': (context) => DragIntoListExample(),
+        '/fixed_example': (context) => FixedExample(),
       },
     );
   }