+++ lib/drag_and_drop_list_interface.dart
@@ -4,9 +4,12 @@ import 'package:drag_and_drop_lists/drag_and_drop_item.dart';
 import 'package:flutter/material.dart';
 
 abstract class DragAndDropListInterface implements DragAndDropInterface {
-  final List<DragAndDropItem> children;
+  List<DragAndDropItem> get children;
 
-  DragAndDropListInterface({this.children});
+  /// Whether or not this item can be dragged.
+  /// Set to true if it can be reordered.
+  /// Set to false if it must remain fixed.
+  bool get canDrag;
 
   Widget generateWidget(DragAndDropBuilderParameters params);
 }